Improve Stock class
Added the functionality to remove the current trading days candlestick from a created stocks TradingPeriod. This was becuase we don't want to add a candlestick before the end of the trading session. This could be removed later maybe in favor for overwriting an already existing candlesteick in the database? Though, this is the Stock class and this class should not now anything about anything else than what is needed. Single Responsibility principle! In addition to these changes, some javadoc was improved in this class.