Add method in Main to use MainDataFetcher for fetching multiple stocks' data
This method was added to Main to compile a JAR used for testing the retrieval of one day's price data over a one-week interval for multiple stocks.
The goal is to determine if fetching data once a day for a week provides accurate and updated volume information, as well as to see if it eliminates the need for multiple fetches.
Previously, the volume data appeared incomplete when fetching only the current day's data, and complete volume data was only received when fetching the following day before the market opened.
This approach aims to reduce API calls to save on costs.